test #29: 관심사 엔티티 중복 테스트 삭제
